fix(LottoOwner, YieldRate): 얻은 금액 타입 수정
* 받은 금액이 정수 타입의 범위를 벗어날 수 있으므로 gainMoney의 타입을 int에서 long으로 변경했다.

+++ src/main/java/lotto/domain/LottoOwner.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class LottoOwner {
     private List<Lotto> lottos;
     private int spendMoney;
-    private int gainMoney;
+    private long gainMoney;
 
     public int purchaseLotto(int money) {
         spendMoney = money;
